4/8/2023 Update
Created a fileManager
It is called whenever the user encrypts a file and writes to the file
Created an option to print the file
In the main menu, printing the file is now option 3, while quitting is option 4.
Made the fileManager able to properly write to the file
The plainText, encryptedText, key, and encryption method are all properly saved to the passwords.txt file.